requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
pytz==2021.3
setuptools==58.5.3
sqlparse==0.4.2
wheel==0.37.0
